src/main/java/com/technical/task/task_project/service/SecurityService.java
package com.technical.task.task_project.service;

import lombok.AllArgsConstructor;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Service;

@AllArgsConstructor
@Service
public class SecurityService {
    private final PasswordEncoder passwordEncoder;

    public String encodePassword(String rawPassword) {
        return passwordEncoder.encode(rawPassword);
    }

    public boolean matchesPassword(String givenPassword, String encodedPassword) {
        return passwordEncoder.matches(givenPassword, encodedPassword);
    }

}
